adding api features and fixing some validation bugs

+++ utils/removeOldImg.js
@@ -1,8 +1,10 @@
 import fs from "fs";
+import { AppError } from "./appError.js";
 export const removeOldImage = function (qualifiedImage) {
-  let startIndex = qualifiedImage.indexOf("up");
-  let oldImagePath = qualifiedImage.slice(startIndex);
-  fs.unlinkSync(oldImagePath);
+  return (req, res, next) => {
+    let startIndex = qualifiedImage.indexOf("up");
+    let oldImagePath = qualifiedImage.slice(startIndex);
+    if (fs.existsSync(oldImagePath)) fs.unlinkSync(oldImagePath);
+    else next(new AppError("file does not exist", 409));
+  };
 };
-
-let them = "ينفعوك";
